# swagger.yaml
openapi: 3.0.0
info:
  title: Node MySQL API
  version: 1.0.0
servers:
  - url: http://localhost:4000
paths:
...
  /accounts/{id}:
    get:
      summary: Get account by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    put:
      summary: Update account
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    delete:
      summary: Delete account
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK